Manage the sort and reverse-sort options.
If one of those options is selected, the result will be appended in a buffer instead of the standard output, and then we will sort the buffer and display it.
In order to avoid sorting human readable sizes, the kilobytes size is always appended as prefix of the line, and after the sort, we remove it, using the cut command.